Move CMS specific files into their own namespace
